+++ packages/subtree-updater/src/cli/index.ts
@@ -14,4 +14,16 @@ export default async function main(): Promise<void> {
   await program.parseAsync(process.argv);
 }
 
-main().catch((e) => console.log(`subtree updater exited with error: ${e}`));
+// ! HACK
+// ! in principle, there should be no hanging promises once `main()` resolves or rejects.
+// ! as a backstop, we manually call `process.exit()` to ensure the process actually exits
+// ! even if there's a bug somewhere that results in a hanging promise
+main()
+  .then(() => {
+    console.log(`subtree-updater-cli ran to completion`);
+    process.exit(0);
+  })
+  .catch((e) => {
+    console.log(`subtree-updater-cli exited with error: ${e}`);
+    process.exit(1);
+  });